Moved binarystring() and mfcycle() macros out of SharedMacros.h
Made them both C functions and simplified them.

+++ Shared/HelperServices/HelperServices.m
@@ -822,7 +822,7 @@ static NSError *makeWritable(NSString *itemPath) {
         
         /// Debug
         
-        DDLogInfo(@"Changed permissions of %@ from %@ to %@", itemPath, binarystring((int)oldPermissions), binarystring((int)newPermissions));
+        DDLogInfo(@"Changed permissions of %@ from %@ to %@", itemPath, binarystring(oldPermissions), binarystring(newPermissions));
         /// ^ Binary representation doesn't really help. This is almost impossible to parse visually.
     }
